Use the detected package manager for RedHat-based platforms
RedHat-based platforms have historically used `yum` as the package
manager but newer releases have switched to preferring `dnf` as the
default package manager. This update respects what Ansible detects as
the package manager to use as this is what will be used by the
`ansible.builtin.package` module unless overridden.