Add code to reinstall the filesystem package if necessary
We have seen on both the Fedora 42 Docker image (fedora:42) and the
Fedora 42 base AMI that we use in AWS that the /usr/local/sbin symlink
to /usr/local/bin does not exist.  In that case one must reinstall the
filesystem package to ensure the directory exists.

Issue #73 exists to remind us to remove this code once it is no longer
needed.